fix crash / exception when trying to load out of range values
The user can load previously saved configuration values - either
automatically as part of the program startup or manually at any
time during the program run. If the loaded values were outside of
the allowed range for those values, this would throw an exception
and potentially crash the application. The problem is fixed by
adjusting the loaded values to the allowed range.

Usually, those configuration values are not out of range. But
since the configuration file can be edited manually, a user can
enter any numeric values into that file, potentially causing the
(now fixed) exception to trigger as soon as the configuration
file is loaded by the application.